fix schema plugin
