+++ eventmesh-registry-plugin/eventmesh-registry-api/build.gradle
@@ -17,6 +17,7 @@
 
 dependencies {
     api project(":eventmesh-spi")
+    implementation project(":eventmesh-common")
 
     testImplementation project(":eventmesh-spi")
 }
